Modify the NSMutableDictionary subclass used internally such that it maintains the order of objects
This is useful for example when using JSONKit to parse JSON that will appear in an editable UI.  In this situation we would like the order in the UI to be identical to that in the file.

We may need to further extend this in future to allow for inserting objects at a certain position (for an editable UI).